+++ docs/doctheme/cjs/partials/footer.d.ts
@@ -1,3 +1,0 @@
-import { JSX } from "typedoc";
-import { DumiThemeRenderContext } from "../contexts/DumiThemeRenderContext";
-export declare function footer(context: DumiThemeRenderContext): JSX.Element;