+++ docs/doctheme/cjs/partials/navigation.d.ts
@@ -1,3 +1,0 @@
-import { JSX, type PageEvent, type Reflection } from 'typedoc';
-import { DumiThemeRenderContext } from '../contexts/DumiThemeRenderContext';
-export declare function navigation(context: DumiThemeRenderContext, event: PageEvent<Reflection>): JSX.Element;