+++ docs/doctheme/cjs/partials/sidebarLinks.d.ts
@@ -1,3 +1,0 @@
-import { JSX } from 'typedoc';
-import { DumiThemeRenderContext } from '../contexts/DumiThemeRenderContext';
-export declare function sidebarLinks(context: DumiThemeRenderContext): JSX.Element | null;